Refactor settings modal
