+++ frontend/src/app/admin/modals.module.css
@@ -12,3 +12,50 @@
   border-radius: 4px;
   height: 40px;
 }
+
+.settingsModal {
+  text-align: center;
+}
+
+.settingsInputContainer {
+  display: grid;
+  grid-template-columns: 10% 50%;
+  padding-bottom: 15px;
+  row-gap: 15px;
+}
+
+.settingsInput {
+  width: 100%;
+  margin: 0 8px;
+  padding: 0 10px;
+  border: 1px solid rgba(var(--foreground-rgb), 0.5);
+  background-color: rgb(var(--background-rgb));
+  border-radius: 4px;
+  height: 40px;
+}
+
+.settingsInputLabel {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.settingsButtonGroup {
+  display: flex;
+  justify-content: space-evenly;
+}
+
+.deleteButton {
+  background-color: rgb(var(--button-rgb));
+  padding: 10px 20px;
+  text-align: center;
+  text-decoration: none;
+  display: inline-block;
+  font-size: 16px;
+  border-radius: 5px;
+  color: black;
+}
+
+.deleteButton:hover {
+  background-color: rgba(255, 0, 0, 0.8);
+}
